feat(ui): organize node utils
