+++ invokeai/frontend/web/src/features/nodes/util/graph/addIPAdapterToLinearGraph.ts
