+++ invokeai/frontend/web/src/features/nodes/util/graph/addNSFWCheckerToGraph.ts
