+++ invokeai/frontend/web/src/features/nodes/util/graph/addSDXLLoRAstoGraph.ts
