+++ invokeai/frontend/web/src/features/nodes/util/graph/addSDXLRefinerToGraph.ts
