+++ invokeai/frontend/web/src/features/nodes/util/graph/addSeamlessToLinearGraph.ts
