+++ invokeai/frontend/web/src/features/nodes/util/graph/addT2IAdapterToLinearGraph.ts
