+++ invokeai/frontend/web/src/features/nodes/util/graph/addVAEToGraph.ts
