+++ invokeai/frontend/web/src/features/nodes/util/graph/addWatermarkerToGraph.ts
