+++ invokeai/frontend/web/src/features/nodes/util/graph/buildAdHocUpscaleGraph.ts
