+++ invokeai/frontend/web/src/features/nodes/util/graph/buildCanvasGraph.ts
