+++ invokeai/frontend/web/src/features/nodes/util/graph/buildCanvasImageToImageGraph.ts
