+++ invokeai/frontend/web/src/features/nodes/util/graph/buildCanvasInpaintGraph.ts
