+++ invokeai/frontend/web/src/features/nodes/util/graph/buildCanvasTextToImageGraph.ts
