+++ invokeai/frontend/web/src/features/nodes/util/graph/buildLinearBatchConfig.ts
