+++ invokeai/frontend/web/src/features/nodes/util/graph/helpers/craftSDXLStylePrompt.ts
