+++ invokeai/frontend/web/src/features/nodes/util/workflow/buildWorkflow.ts
@@ -1,6 +1,6 @@
 import { logger } from 'app/logging/logger';
-import { NodesState } from '../store/types';
-import { WorkflowV2, zWorkflowEdge, zWorkflowNode } from '../types/workflow';
+import { NodesState } from '../../store/types';
+import { WorkflowV2, zWorkflowEdge, zWorkflowNode } from '../../types/workflow';
 import { fromZodError } from 'zod-validation-error';
 import { parseify } from 'common/util/serialize';
 import i18n from 'i18next';
